Cliente nuevos,compra noficaiones
-Agregar acceso directo  para editar informacion de pacientes nuevos y
especificar la informacion incompleta
-Agregar notificaciones al guardar o preguardar la compra.

+++ src/hot/com/sa/kubekit/action/inventory/CompraHome.java
@@ -137,7 +137,10 @@ public class CompraHome extends KubeDAO<Compra>{
 	public void removerItemPreCompra(Item item){
 		itemsAgregados.remove(item);
 		productosAgregados.remove(item.getInventario());
-		getEntityManager().remove(item);
+		
+		if(item.getItemId().getMovimientoId()!=null && item.getItemId().getInventarioId()!=null)
+			getEntityManager().remove(item);
+		
 		actualizarSubtotal();
 	}
 	
@@ -454,48 +457,53 @@ public void cargarListaCodigos(ItemPedido prdItm) {
 	{
 		
 		
-		System.out.println("Cantidad"+item.getCantidad());
-		System.out.println("Nombre"+item.getInventario().getProducto().getNombre());
-		
-		
-		//Solucion
-		//Compara con la categoria del inventario si este requiere serie o lote y cargar la lista de codigos que tiene con el metodo existente
-		
-		if(item.getInventario().getProducto().getCategoria().isTieneNumLote() || item.getInventario().getProducto().getCategoria().isTieneNumSerie())
+			System.out.println("Cantidad"+item.getCantidad());
+			System.out.println("Nombre"+item.getInventario().getProducto().getNombre());
+			
+		if(item.getItemId().getMovimientoId()!=null && item.getItemId().getInventarioId()!=null)
 		{
+			//Solucion
+			//Compara con la categoria del inventario si este requiere serie o lote y cargar la lista de codigos que tiene con el metodo existente
 			
-			//recorrer la lista de codigos y eliminar o desactivar los codigos
-			cargarListaCodigos(item);
-			
-			if(lstCodsProductos.get(item.getInventario().getProducto().getReferencia())!=null && lstCodsProductos.get(item.getInventario().getProducto().getReferencia()).get(0).getId()!=null)
+			if(item.getInventario().getProducto().getCategoria().isTieneNumLote() || item.getInventario().getProducto().getCategoria().isTieneNumSerie())
 			{
-				for(CodProducto codigo: lstCodsProductos.get(item.getInventario().getProducto().getReferencia()))
+				
+				//recorrer la lista de codigos y eliminar o desactivar los codigos
+				cargarListaCodigos(item);
+				
+				if(lstCodsProductos.get(item.getInventario().getProducto().getReferencia())!=null && lstCodsProductos.get(item.getInventario().getProducto().getReferencia()).get(0).getId()!=null)
 				{
-					codigo.setEstado("USD");
-					getEntityManager().merge(codigo);
+					for(CodProducto codigo: lstCodsProductos.get(item.getInventario().getProducto().getReferencia()))
+					{
+						codigo.setEstado("USD");
+						getEntityManager().merge(codigo);
+					}
+					
+				}
+				else
+				{
+					System.out.println("Lista de codigos null");
 				}
 				
 			}
-			else
-			{
-				System.out.println("Lista de codigos null");
-			}
 			
+			Movimiento movimiento = new Movimiento();
+			movimiento.setRazon("O");
+			movimiento.setSucursal(item.getInventario().getSucursal());
+			movimiento.setTipoMovimiento("S");
+			
+			
+			removerItemPreCompra(item);
+			
+			movimientoHome.getItemsAgregados().clear();
+			movimientoHome.setInstance(movimiento);
+			movimientoHome.getItemsAgregados().add(item);
+			movimientoHome.save();
 		}
-		
-		Movimiento movimiento = new Movimiento();
-		movimiento.setRazon("O");
-		movimiento.setSucursal(item.getInventario().getSucursal());
-		movimiento.setTipoMovimiento("S");
-		
-		
-		removerItemPreCompra(item);
-		
-		movimientoHome.setInstance(movimiento);
-		movimientoHome.getItemsAgregados().add(item);
-		movimientoHome.save();
-		
-		
+		else
+		{
+			removerItemPreCompra(item);
+		}
 		
 		
 	}
